mirror/backend/requirements.txt
openai
python-dotenv
eth-account
requests
streamlit
